+++ src/peekableView.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export const ScreenContext = React.createContext();
+
+export default View;
